调整 drawer 样式

+++ scripts/watch.js
@@ -13,14 +13,17 @@ chokidar.watch(srcPath, {
   if (file.endsWith('.ts') && !file.endsWith('.d.ts')) return;
   if (file.endsWith('.tsx')) return;
   let path = file.split(srcPath)[1];
-  console.log(event, path);
   let target = Path.join('node_modules/@samoyed', path);
   switch (event) {
     case 'modified':
+      if (fs.existsSync(target) && fs.readFileSync(file).compare(fs.readFileSync(target)) === 0) {
+        return;
+      }
       fs.copyFileSync(file, target);
       break;
     case 'moved':
       del.sync(target);
       break;
   }
+  console.log(event, path);
 });
